Do not delete configs on a readonly mount
Following a recent change to the jenkins deployment that made the
configmap and the secrets readonly, deleting the configurations that
are stored in the config-map resulted in an error message.
This patch fixes it by deleting `rm -rf /opt/openshift/configuration`
in the `run` script.

Fixes: https://github.com/openshiftio/openshift.io/issues/2608
See also: https://github.com/kubernetes/kubernetes/issues/60814