Save space on gluster storage by uncompressing the plugins in image
We want to save some space on gluster mount and we want to use more on the image
instead which is volatile.

In a nutshell this is what we do :

- We uncompress the jpi/hpi in the image directly.
- We symlinks the hpi and jpi into /var/lib/jenkins/plugins at (*every*) boot time
- We symlinks the /var/lib/jenkins/plugins/${PLUGIN_NAME}/WEB-INF
- We let jenkins do its magic for the other files.

Now there is one caveat here, to let it does its magic, jenkins does a bunch of
things to uncompress and update stuff and that's not a behavior that it seems
possible to disable.

So we let jenkins do that at the first boot which will have a 215mb for the
jenkins plugins. On the next reboot we will overwrite all of the WEB-INF (which
is the dirs will the contents) with the one from the image.

Jenkins is fooled and we are down 29mb for /var/lib/jenkins/plugins

Since we are using plugins from the image we should be okay with updates (to be
checked).